requirements.txt
flask
pycryptodome



